Compliance: Moves add- and amend_fields to compliance_check
* Due to missing generality, the functions add_fields and
  amend_fields are moved from utils.py to compliance_check.py
