+++ compliance_check.py
@@ -18,7 +18,24 @@
 ## 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
 
 from invenio.NonComplianceCheck import NonComplianceChecks
-from invenio.utils import amend_fields, add_fields
+
+
+def amend_fields(record, field, changes_dict, comment=''):
+    """record : AmendableRecord"""
+    for position, value in record.iterfield(field):
+        key = value[:value.find(':')]
+        if key in changes_dict:
+            record.amend_field(position,
+                               '%s:%d' % (key, changes_dict[key]),
+                               comment)
+
+
+def add_fields(record, field, subfield, value_dict):
+    """record : AmendableRecord"""
+    for key, value in value_dict.iteritems():
+        record.add_field(field,
+                         value="",
+                         subfields=[(subfield, '%s:%d' % (key, value))])
 
 
 def check_records(records, empty=False):
